LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-in-attribute-expected.txt

CONSOLE MESSAGE: line 26: SyntaxError: Unexpected token ')'

PASS window.onerror - compile error in attribute 
PASS window.onerror - compile error in attribute (column) 
